modelisation I5 1718

+++ ipeda/src/main/java/fr/epsi/ipeda/service/database/IDatabaseService.java
@@ -24,17 +24,17 @@ public interface IDatabaseService {
 
 	public void getPlanningBySemaine(TypeFormation typeFormation, Semaine semaine);
 
-	Module createModule(String codeModule, String libelle, int dureeFFP, int dureeTE, Map<NumeroSemestre, Semestre> mapSemestres, Intervenant intervenant, UniteEnseignement ue);
+	Module createModule(String codeModule, String libelle, double dureeFFP, double dureeTE, Map<NumeroSemestre, Semestre> mapSemestres, Intervenant intervenant, UniteEnseignement ue);
 
 	void createModuleCompose(Module moduleParent, Module... modules);
 
-	ProjetTransversal createProjetTransversal(String codeModule, String libelle, int dureeFFP, int dureeTE, Map<NumeroSemestre, Semestre> mapSemestres, Intervenant intervenant,
+	ProjetTransversal createProjetTransversal(String codeModule, String libelle, double dureeFFP, double dureeTE, Map<NumeroSemestre, Semestre> mapSemestres, Intervenant intervenant,
 			BlocCompetences blocCompetences);
 
-	void createCours(String dateHeureDebut, int duree, String codeModule, CodeSalle codeSalle);
+	void createCours(String dateHeureDebut, double duree, String codeModule, CodeSalle codeSalle);
 
 	void createModuleMutualise(String libelle, ArrayList<Module> listeModulesMutualises);
 
-	public Module createModule(Module moduleParent, String libelle, int dureeFFP, int dureeTE, Intervenant intervenant);
+	public Module createModule(Module moduleParent, String libelle, double dureeFFP, double dureeTE, Intervenant intervenant);
 
 }
